loop/requirements.txt
requests>=2.31.0
anthropic>=0.30.0
pyyaml>=6.0.1

qdrant-client>=1.7.0
openai>=1.0.0
